refactor(app): group resource tRPC routers under resources/
Group the UI-facing resource API into trpc/resources/ (agent, agent-type,
skill, template, shared-env-set, plus index.ts and shared.ts, which are
group-private: the SuperJSON initTRPC and Better Auth context are used
only by these routers). telemetry.ts stays at trpc/ root as the separate
plugin-facing API surface with its own transformer-free initTRPC.

Update the three external import sites (router.tsx, server.ts,
server.test.ts); internal relative imports are unchanged since index.ts
and shared.ts moved together with the routers.
