+++ apps/app/src/server/routers/trpc/resources/template.ts
